+++ src/esm/utils/Params.js
@@ -6,6 +6,7 @@
  * @version  1.0.0
  * @date     2023-03-13
  * @modified 2024-08-26 Added the `hasParam` method.
+ * @modified 2026-05-22 Added the `isValueAllowed` function param to the getter methods.
  */
 export class Params {
     constructor(baseParams) {
@@ -14,19 +15,26 @@ export class Params {
     hasParam(name) {
         return this.baseParams.hasOwnProperty(name);
     }
-    getString(name, fallback) {
+    getString(name, fallback, isValueAllowed) {
         let value = this.baseParams[name];
         if (typeof value === "undefined" || !value || (value = value.trim()).length === 0) {
+            return fallback;
+        }
+        if (isValueAllowed && !isValueAllowed(value)) {
             return fallback;
         }
         return value;
     }
-    getNumber(name, fallback) {
+    getNumber(name, fallback, isValueAllowed) {
         let value = this.baseParams[name];
         if (typeof value === "undefined" || !value || (value = value.trim()).length === 0) {
             return fallback;
         }
-        return Number(value);
+        var numericValue = Number(value);
+        if (isValueAllowed && !isValueAllowed(numericValue)) {
+            return fallback;
+        }
+        return numericValue;
     }
     getBoolean(name, fallback) {
         let value = this.baseParams[name];